Split eeprom code into separate module
