+++ src/main/io/serial_msp.c
@@ -91,6 +91,7 @@
 #include "fc/runtime_config.h"
 
 #include "config/config.h"
+#include "config/config_eeprom.h"
 #include "config/config_profile.h"
 #include "config/config_master.h"
